pyproject.toml
[tool.poetry]
name = "certainty-py"
version = "0.1.0"
description = ""
authors = ["Joseph Redfern <joseph@redfern.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.2"
rq = "^1.16.2"
tortoise-orm = "^0.21.6"
uvicorn = "^0.30.6"
rq-scheduler = "^0.13.1"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
starsessions = "^2.1.3"
sendgrid = "^6.11.0"
python-dotenv = "^1.0.1"
validators = "^0.34.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
